Explicitly handle SIGRTMIN+3 signal.
SIGRTMIN+3 is a stop signal for resin base images to gracefully stop
systemd when it is enabled. However, the stop signal should be SIGTERM
in normal cases so this PR will change how TINI handles SIGRTMIN+3.

When Tini receives SIGRTMIN+3, SIGTERM will be forwarded to child
process as it receives SIGTERM.